UI Header popover cleanup, rough settings page
